+++ exp/source/frontend/parser.c
@@ -658,7 +658,7 @@ static bool identifier(Operand *result, Parser *parser, Context *context) {
         return error(parser, context, ERROR_TYPECHECK_UNDEFINED_SYMBOL);
     }
 
-    u16 index = context_labels_insert(context, name);
+    u32 index = context_labels_insert(context, name);
     *result   = operand_label(index);
     return true;
 }
